Homepage update
Now the homepage is visible to every user, anonymous or logged. Every tipe of user will see different links on the navbar located on the top of the webpage.

The anonymous user will see only login ad register links
The user, instructor and manager will continue to see the same links that they can find on the rest of their logged area